+++ src/components/MarkerListAccordion.js
@@ -1,16 +1,60 @@
+/**
+ * ${SUMMARY}
+ * 
+ * ${DESCRIPTION}
+ * 
+ * @author  chris-m92
+ * 
+ * MIT License
+ * 
+ * Copyright (c) 2020 Airmen Coders
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+//----------------------------------------------------------------//
+// Top Level Modules
+//----------------------------------------------------------------//
 import React from 'react'
+import { LatLon } from 'geodesy/mgrs'
 
+//----------------------------------------------------------------//
+// Material-UI Components
+//----------------------------------------------------------------//
 import Accordion from '@material-ui/core/Accordion'
 import AccordionDetails from '@material-ui/core/AccordionDetails'
 import AccordionSummary from '@material-ui/core/AccordionSummary'
 import { makeStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
 
-import { LatLon } from 'geodesy/mgrs'
-
+//----------------------------------------------------------------//
+// Material-UI Icons
+//----------------------------------------------------------------//
 import AttachFileIcon from '@material-ui/icons/AttachFile'
+
+//----------------------------------------------------------------//
+// Custom Functions
+//----------------------------------------------------------------//
 import { render9line, render15line } from '../functions/renderData'
 
+//----------------------------------------------------------------//
+// Custom Styling
+//----------------------------------------------------------------//
 const useStyles = makeStyles(theme => ({
   heading: {
     fontSize: theme.typography.pxToRem(15),
@@ -23,6 +67,9 @@ const useStyles = makeStyles(theme => ({
   },
 }))
 
+//----------------------------------------------------------------//
+// Marker List Accordion Component
+//----------------------------------------------------------------//
 const MarkerListAccordion = (props) => {
 
   const classes = useStyles()